tests/wpt/web-platform-tests/dom/ranges/Range-getBoundingClientRect.html

<!DOCTYPE html>
<title>Range getBoundingClientRect</title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<div id=data>Testing data
  <p>Paragraph</p>
</div>
<div id=data2>Data 2
  <p id=data3>Data 3</p>
</div>
<script>
test(function() {
  var values = {x: 8, y: 28.800003051757812, width: 1280,
    height: 123.19999694824219, top: 28.800003051757812, right: 1288,
    bottom: 152, left: 8}
  var r = document.createRange();
  r.setStart(document.getElementById('data'), 0);
  r.setEnd(document.getElementById('data3'), 0);

  var rect = r.getBoundingClientRect();
  console.log(rect);
  for (var property in values) {
    assert_equals(values[property], rect[property]);
  }
})
</script>
